Add test script to set up ofono-phonesim
Add tests/with-ofono-phonesim. Running this as root will set up ofono
to run with the phone simulator. This uses an unshared tmpfs mount over
/etc/ofono/ to avoid having to modify the file system.

The script can either be invoked without arguments and run until Control-C,
or run a program (with arguments), in which case everything will be
restored when the program ends.